feat(patch): v2.0.1 polish — health check, loading indicators, handler split, monarch tests
- feat(web): add GET /health endpoint with db_ok, lnd_configured, version, uptime (#101)
- feat(web): add HTMX global progress bar; hx-disabled-elt on import buttons (#105)
- refactor(web): split handler_html.go into per-page files (dashboard, lightning, partials,
  imports, pl, tax, wallets, settings, transactions, portfolio) (#109)
- test(monarch): add SyncTransactions, merchantForTx, sourceLabel, ensureDefaultCategory
  tests with mock TransactionClient and CategoryLister (#108)
- chore(docker): add HEALTHCHECK directive using /health endpoint
