+++ internal/web/handler_html.go
@@ -1,2480 +1,1 @@
 package web
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"net/http"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/satsbook/satsbook/internal/db"
-	"github.com/satsbook/satsbook/internal/license"
-	"github.com/satsbook/satsbook/internal/lnd"
-	"github.com/satsbook/satsbook/internal/monarch"
-	"github.com/satsbook/satsbook/internal/tax"
-)
-
-// DashboardData holds all data needed to render the dashboard page.
-type DashboardData struct {
-	// License tier (free, pro, power)
-	Tier string
-
-	// Node info
-	NodeAlias  string
-	NodePubKey string
-	NodeSynced bool
-	BlockHeight uint32
-	LNDVersion string
-
-	// Fee summary
-	FeesAllTimeSats int64
-	FeesAllTimeUSD  float64
-	Fees30dSats     int64
-	Fees7dSats      int64
-
-	// Routed counts
-	RoutedAllTime int64
-	Routed30d     int64
-	Routed7d      int64
-
-	// Other stats
-	ActiveChannels          int
-	WalletBalanceSats       int64
-	WalletBalanceUSD        float64
-	ChannelLocalBalanceSats int64
-	ChannelLocalBalanceUSD  float64
-
-	// Exchange balances
-	StrikeBalanceSats   int64
-	StrikeBalanceUSD    float64
-	RiverBalanceSats      int64
-	RiverBalanceUSD       float64
-	CoinbaseBalanceSats   int64
-	CoinbaseBalanceUSD    float64
-	SwanBalanceSats       int64
-	SwanBalanceUSD        float64
-	ExchangeBalanceSats   int64
-	ExchangeBalanceUSD    float64
-
-	// Cost basis (from exchange imports)
-	TotalCostBasisUSD float64
-	AvgCostPerBTC     float64 // TotalCostBasisUSD / total purchased BTC
-
-	// Cold storage (watched wallets)
-	ColdStorageSats int64
-	ColdStorageUSD  float64
-
-	// Headline: total BTC under control (wallet + channels + exchange + cold storage)
-	TotalBTCSats int64
-	TotalBTCUSD  float64
-
-	// YTD section
-	YTDRoutingFeesSats int64
-	YTDRoutingFeesUSD  float64
-	YTDPurchasedSats   int64
-	YTDPurchasedUSD    float64
-	YTDNetChangeSats   int64
-	YTDNetChangeUSD    float64
-
-	// Onboarding state flags
-	LNDConnected       bool
-	HasFeeHistory      bool
-	HasExchangeImports bool
-	ShowOnboarding     bool // !HasFeeHistory && !HasExchangeImports
-	ShowImportBanner   bool // HasFeeHistory && !HasExchangeImports
-	ShowLNDBanner      bool // HasExchangeImports && !LNDConnected
-
-	// Price
-	BTCPriceUSD    float64
-	PriceFetchedAt time.Time
-
-	// Sync
-	LastSyncedAt time.Time
-
-	// Chart data
-	DailyFees          []db.DailyFeeStat
-	PortfolioSnapshots []db.PortfolioSnapshot
-
-	// Channel list
-	Channels []db.ChannelStat
-
-	// Forwarding defaults
-	DefaultFrom string
-	DefaultTo   string
-}
-
-// ForwardingTableData holds data for the forwarding events partial.
-type ForwardingTableData struct {
-	Events    []db.ForwardingEvent
-	Page      int
-	Limit     int
-	Total     int64
-	PageStart int
-	PageEnd   int
-	HasNext   bool
-	From      string
-	To        string
-}
-
-// HandleDashboard serves GET / (main dashboard page).
-func (h *Handler) HandleDashboard(w http.ResponseWriter, r *http.Request) {
-	if r.URL.Path != "/" {
-		http.NotFound(w, r)
-		return
-	}
-
-	ctx := r.Context()
-	now := time.Now()
-	since30d := now.AddDate(0, 0, -30)
-	since7d := now.AddDate(0, 0, -7)
-	sinceYTD := time.Date(now.Year(), 1, 1, 0, 0, 0, 0, time.UTC)
-	data := DashboardData{
-		Tier:        string(TierFromContext(ctx)),
-		DefaultFrom: since30d.Format("2006-01-02"),
-		DefaultTo:   now.Format("2006-01-02"),
-	}
-
-	// Fetch all data concurrently for the main dashboard
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	type result struct {
-		fees30d, routed30d int64
-		fees7d, routed7d   int64
-		activeChannels     int
-		balance            *db.WalletBalanceSnapshot
-		channels           []db.ChannelStat
-		dailyFees          []db.DailyFeeStat
-		nodeInfo           *lnd.NodeInfo
-		lastSynced         time.Time
-		btcPrice           float64
-		priceFetched       time.Time
-		portfolioAll       *db.PortfolioPositionResult
-		portfolioYTD       *db.PortfolioPositionResult
-		coldStorageSats int64
-	}
-	var res result
-
-	fetch := func(fn func()) {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			fn()
-		}()
-	}
-
-	fetch(func() {
-		// PortfolioPosition replaces the previous all-time FeeSummary + 3x ExchangeBalance
-		// calls with one query for exchange totals + one for routing fees.
-		p, err := h.store.PortfolioPosition(ctx, time.Time{})
-		if err == nil {
-			mu.Lock()
-			res.portfolioAll = p
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		p, err := h.store.PortfolioPosition(ctx, sinceYTD)
-		if err == nil {
-			mu.Lock()
-			res.portfolioYTD = p
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		fees30d, routed30d, err := h.store.FeeSummary(ctx, since30d)
-		if err == nil {
-			mu.Lock()
-			res.fees30d = fees30d
-			res.routed30d = routed30d
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		fees7d, routed7d, err := h.store.FeeSummary(ctx, since7d)
-		if err == nil {
-			mu.Lock()
-			res.fees7d = fees7d
-			res.routed7d = routed7d
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		count, err := h.store.ActiveChannelCount(ctx)
-		if err == nil {
-			mu.Lock()
-			res.activeChannels = count
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		bal, err := h.store.LatestWalletBalance(ctx)
-		if err == nil {
-			mu.Lock()
-			res.balance = bal
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		stats, err := h.store.ChannelStats(ctx)
-		if err == nil {
-			mu.Lock()
-			res.channels = stats
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		fees, err := h.store.DailyFees(ctx, since30d)
-		if err == nil {
-			mu.Lock()
-			res.dailyFees = fees
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		info := h.getNodeInfo(ctx)
-		if info != nil {
-			mu.Lock()
-			res.nodeInfo = info
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		t, err := h.store.LastSyncedAt(ctx)
-		if err == nil {
-			mu.Lock()
-			res.lastSynced = t
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		price, err := h.price.GetBTCPrice(ctx)
-		if err == nil {
-			mu.Lock()
-			res.btcPrice = price
-			res.priceFetched = h.price.FetchedAt()
-			mu.Unlock()
-		}
-	})
-
-	if h.walletStore != nil {
-		fetch(func() {
-			total, err := h.walletStore.TotalWatchedBalance(ctx)
-			if err == nil {
-				mu.Lock()
-				res.coldStorageSats = total
-				mu.Unlock()
-			}
-		})
-	}
-
-	wg.Wait()
-
-	// Assemble dashboard template data
-	data.ColdStorageSats = res.coldStorageSats
-
-	if res.nodeInfo != nil {
-		data.NodeAlias = res.nodeInfo.Alias
-		data.NodePubKey = res.nodeInfo.PubKey
-		data.NodeSynced = res.nodeInfo.Synced
-		data.BlockHeight = res.nodeInfo.BlockHeight
-		data.LNDVersion = res.nodeInfo.Version
-	}
-
-	// All-time fees + per-source balances come from PortfolioPosition.
-	if res.portfolioAll != nil {
-		data.FeesAllTimeSats = res.portfolioAll.RoutingFeesSats
-		data.RoutedAllTime = res.portfolioAll.RoutedCount
-		data.StrikeBalanceSats = res.portfolioAll.BySource["strike"].NetSats
-		data.RiverBalanceSats = res.portfolioAll.BySource["river"].NetSats
-		data.CoinbaseBalanceSats = res.portfolioAll.BySource["coinbase"].NetSats
-		data.SwanBalanceSats = res.portfolioAll.BySource["swan"].NetSats
-		data.ExchangeBalanceSats = res.portfolioAll.ExchangeNetSats
-		data.TotalCostBasisUSD = res.portfolioAll.TotalCostBasisUSD
-		if res.portfolioAll.PurchasedSats > 0 {
-			data.AvgCostPerBTC = res.portfolioAll.TotalCostBasisUSD / (float64(res.portfolioAll.PurchasedSats) / 1e8)
-		}
-	}
-
-	data.Fees30dSats = res.fees30d / 1000
-	data.Fees7dSats = res.fees7d / 1000
-	data.Routed30d = res.routed30d
-	data.Routed7d = res.routed7d
-	data.ActiveChannels = res.activeChannels
-	data.LastSyncedAt = res.lastSynced
-	data.DailyFees = res.dailyFees
-	data.Channels = res.channels
-
-	if res.balance != nil {
-		data.WalletBalanceSats = res.balance.TotalSat
-	}
-
-	// Sum channel local balances
-	for _, ch := range res.channels {
-		data.ChannelLocalBalanceSats += ch.LocalBalance
-	}
-
-	// Headline: total BTC = wallet + channels + exchange + cold storage (routing fees already in channels)
-	data.TotalBTCSats = data.WalletBalanceSats + data.ChannelLocalBalanceSats + data.ExchangeBalanceSats + data.ColdStorageSats
-
-	// YTD section
-	if res.portfolioYTD != nil {
-		data.YTDRoutingFeesSats = res.portfolioYTD.RoutingFeesSats
-		data.YTDPurchasedSats = res.portfolioYTD.PurchasedSats
-		data.YTDNetChangeSats = res.portfolioYTD.RoutingFeesSats + res.portfolioYTD.ExchangeNetSats
-	}
-
-	// Onboarding flags
-	data.LNDConnected = res.nodeInfo != nil
-	data.HasFeeHistory = data.FeesAllTimeSats > 0
-	data.HasExchangeImports = data.ExchangeBalanceSats != 0 ||
-		data.StrikeBalanceSats != 0 || data.RiverBalanceSats != 0 || data.CoinbaseBalanceSats != 0 || data.SwanBalanceSats != 0
-	data.ShowOnboarding = !data.HasFeeHistory && !data.HasExchangeImports
-	data.ShowImportBanner = data.HasFeeHistory && !data.HasExchangeImports
-	data.ShowLNDBanner = data.HasExchangeImports && !data.LNDConnected
-
-	if res.btcPrice > 0 {
-		data.BTCPriceUSD = res.btcPrice
-		data.PriceFetchedAt = res.priceFetched
-		data.FeesAllTimeUSD = satsToUSD(data.FeesAllTimeSats, res.btcPrice)
-		data.WalletBalanceUSD = satsToUSD(data.WalletBalanceSats, res.btcPrice)
-		data.ChannelLocalBalanceUSD = satsToUSD(data.ChannelLocalBalanceSats, res.btcPrice)
-		data.ColdStorageUSD = satsToUSD(data.ColdStorageSats, res.btcPrice)
-		data.StrikeBalanceUSD = satsToUSD(data.StrikeBalanceSats, res.btcPrice)
-		data.RiverBalanceUSD = satsToUSD(data.RiverBalanceSats, res.btcPrice)
-		data.CoinbaseBalanceUSD = satsToUSD(data.CoinbaseBalanceSats, res.btcPrice)
-		data.SwanBalanceUSD = satsToUSD(data.SwanBalanceSats, res.btcPrice)
-		data.ExchangeBalanceUSD = satsToUSD(data.ExchangeBalanceSats, res.btcPrice)
-		data.TotalBTCUSD = satsToUSD(data.TotalBTCSats, res.btcPrice)
-		data.YTDRoutingFeesUSD = satsToUSD(data.YTDRoutingFeesSats, res.btcPrice)
-		data.YTDPurchasedUSD = satsToUSD(data.YTDPurchasedSats, res.btcPrice)
-		data.YTDNetChangeUSD = satsToUSD(data.YTDNetChangeSats, res.btcPrice)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "layout", data); err != nil {
-		h.logger.Printf("failed to render dashboard: %v", err)
-	}
-}
-
-// HandleLightningPage serves GET /lightning (Lightning node stats page).
-func (h *Handler) HandleLightningPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	now := time.Now()
-	since30d := now.AddDate(0, 0, -30)
-	since7d := now.AddDate(0, 0, -7)
-	data := DashboardData{
-		DefaultFrom: since30d.Format("2006-01-02"),
-		DefaultTo:   now.Format("2006-01-02"),
-	}
-
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	type result struct {
-		feesAllTime, routedAllTime int64
-		fees30d, routed30d         int64
-		fees7d, routed7d           int64
-		activeChannels             int
-		balance                    *db.WalletBalanceSnapshot
-		channels                   []db.ChannelStat
-		dailyFees                  []db.DailyFeeStat
-		nodeInfo                   *lnd.NodeInfo
-		lastSynced                 time.Time
-		btcPrice                   float64
-		priceFetched               time.Time
-	}
-	var res result
-
-	fetch := func(fn func()) {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			fn()
-		}()
-	}
-
-	fetch(func() {
-		fees, routed, err := h.store.FeeSummary(ctx, time.Time{})
-		if err == nil {
-			mu.Lock()
-			res.feesAllTime = fees
-			res.routedAllTime = routed
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		fees, routed, err := h.store.FeeSummary(ctx, since30d)
-		if err == nil {
-			mu.Lock()
-			res.fees30d = fees
-			res.routed30d = routed
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		fees, routed, err := h.store.FeeSummary(ctx, since7d)
-		if err == nil {
-			mu.Lock()
-			res.fees7d = fees
-			res.routed7d = routed
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		count, err := h.store.ActiveChannelCount(ctx)
-		if err == nil {
-			mu.Lock()
-			res.activeChannels = count
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		bal, err := h.store.LatestWalletBalance(ctx)
-		if err == nil {
-			mu.Lock()
-			res.balance = bal
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		stats, err := h.store.ChannelStats(ctx)
-		if err == nil {
-			mu.Lock()
-			res.channels = stats
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		fees, err := h.store.DailyFees(ctx, since30d)
-		if err == nil {
-			mu.Lock()
-			res.dailyFees = fees
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		info := h.getNodeInfo(ctx)
-		if info != nil {
-			mu.Lock()
-			res.nodeInfo = info
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		t, err := h.store.LastSyncedAt(ctx)
-		if err == nil {
-			mu.Lock()
-			res.lastSynced = t
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		price, err := h.price.GetBTCPrice(ctx)
-		if err == nil {
-			mu.Lock()
-			res.btcPrice = price
-			res.priceFetched = h.price.FetchedAt()
-			mu.Unlock()
-		}
-	})
-
-	wg.Wait()
-
-	if res.nodeInfo != nil {
-		data.NodeAlias = res.nodeInfo.Alias
-		data.NodePubKey = res.nodeInfo.PubKey
-		data.NodeSynced = res.nodeInfo.Synced
-		data.BlockHeight = res.nodeInfo.BlockHeight
-		data.LNDVersion = res.nodeInfo.Version
-	}
-
-	data.FeesAllTimeSats = res.feesAllTime / 1000
-	data.RoutedAllTime = res.routedAllTime
-	data.Fees30dSats = res.fees30d / 1000
-	data.Fees7dSats = res.fees7d / 1000
-	data.Routed30d = res.routed30d
-	data.Routed7d = res.routed7d
-	data.ActiveChannels = res.activeChannels
-	data.LastSyncedAt = res.lastSynced
-	data.DailyFees = res.dailyFees
-	data.Channels = res.channels
-
-	if res.balance != nil {
-		data.WalletBalanceSats = res.balance.TotalSat
-	}
-
-	if res.btcPrice > 0 {
-		data.BTCPriceUSD = res.btcPrice
-		data.PriceFetchedAt = res.priceFetched
-		data.FeesAllTimeUSD = satsToUSD(data.FeesAllTimeSats, res.btcPrice)
-		data.WalletBalanceUSD = satsToUSD(data.WalletBalanceSats, res.btcPrice)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "lightning_layout", data); err != nil {
-		h.logger.Printf("failed to render lightning page: %v", err)
-	}
-}
-
-// HandlePortfolioChartPartial serves GET /partials/portfolio-chart (HTMX partial).
-func (h *Handler) HandlePortfolioChartPartial(w http.ResponseWriter, r *http.Request) {
-	days := parseIntParam(r, "days", 30)
-	if days <= 0 {
-		days = 3650 // "All" — 10 years
-	}
-
-	snaps, err := h.store.PortfolioSnapshots(r.Context(), days)
-	if err != nil {
-		h.logger.Printf("portfolio chart partial error: %v", err)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	// portfolioChart returns template.HTML, write it directly
-	html := portfolioChart(snaps)
-	w.Write([]byte(html))
-}
-
-// HandleForwardingPartial serves GET /partials/forwarding (HTMX partial).
-func (h *Handler) HandleForwardingPartial(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	now := time.Now()
-	from, err := parseDateParam(r, "from", now.AddDate(0, 0, -30))
-	if err != nil {
-		h.writeError(w, http.StatusBadRequest, "invalid 'from' date")
-		return
-	}
-	to, err := parseDateParam(r, "to", now)
-	if err != nil {
-		h.writeError(w, http.StatusBadRequest, "invalid 'to' date")
-		return
-	}
-	// Extend 'to' to end of day
-	to = to.Add(24*time.Hour - time.Second)
-
-	page := parseIntParam(r, "page", 1)
-	if page < 1 {
-		page = 1
-	}
-	limit := parseIntParam(r, "limit", 20)
-	if limit < 1 {
-		limit = 1
-	}
-	if limit > 100 {
-		limit = 100
-	}
-
-	offset := (page - 1) * limit
-
-	result, err := h.store.ForwardingEvents(ctx, from, to, limit, offset)
-	if err != nil {
-		h.logger.Printf("failed to query forwarding events: %v", err)
-		http.Error(w, "failed to load forwarding events", http.StatusInternalServerError)
-		return
-	}
-
-	pageStart := offset + 1
-	if result.Total == 0 {
-		pageStart = 0
-	}
-	pageEnd := offset + len(result.Events)
-
-	data := ForwardingTableData{
-		Events:    result.Events,
-		Page:      page,
-		Limit:     limit,
-		Total:     result.Total,
-		PageStart: pageStart,
-		PageEnd:   pageEnd,
-		HasNext:   int64(pageEnd) < result.Total,
-		From:      r.URL.Query().Get("from"),
-		To:        r.URL.Query().Get("to"),
-	}
-	if data.From == "" {
-		data.From = now.AddDate(0, 0, -30).Format("2006-01-02")
-	}
-	if data.To == "" {
-		data.To = now.Format("2006-01-02")
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "forwarding_table", data); err != nil {
-		h.logger.Printf("failed to render forwarding partial: %v", err)
-	}
-}
-
-// ImportPageData is passed to the import page template.
-type ImportPageData struct {
-	HasStrike   bool
-	HasRiver    bool
-	HasCoinbase bool
-	HasSwan     bool
-}
-
-// HandleImportPage serves GET /import.
-func (h *Handler) HandleImportPage(w http.ResponseWriter, r *http.Request) {
-	data := ImportPageData{}
-	// Query per-source presence. We only care whether the vendor has any
-	// rows so the danger zone can be revealed; any error here is non-fatal.
-	if pos, err := h.store.PortfolioPosition(r.Context(), time.Time{}); err == nil && pos != nil {
-		if s, ok := pos.BySource["strike"]; ok && (s.NetSats != 0 || s.PurchasedSats != 0) {
-			data.HasStrike = true
-		}
-		if s, ok := pos.BySource["river"]; ok && (s.NetSats != 0 || s.PurchasedSats != 0) {
-			data.HasRiver = true
-		}
-		if s, ok := pos.BySource["coinbase"]; ok && (s.NetSats != 0 || s.PurchasedSats != 0) {
-			data.HasCoinbase = true
-		}
-		if s, ok := pos.BySource["swan"]; ok && (s.NetSats != 0 || s.PurchasedSats != 0) {
-			data.HasSwan = true
-		}
-	} else if err != nil {
-		h.logger.Printf("import page: portfolio position lookup failed: %v", err)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "import_layout", data); err != nil {
-		h.logger.Printf("failed to render import page: %v", err)
-	}
-}
-
-// PeriodOption represents a selectable time period for the P&L page.
-type PeriodOption struct {
-	Value  string
-	Label  string
-	Active bool
-}
-
-// PLPageData holds all data needed to render the P&L page.
-type PLPageData struct {
-	// Layout fields (required by pl_layout.html)
-	NodeAlias      string
-	NodePubKey     string
-	NodeSynced     bool
-	BlockHeight    uint32
-	LNDVersion     string
-	LastSyncedAt   time.Time
-	BTCPriceUSD    float64
-	PriceFetchedAt time.Time
-
-	// Period selector
-	SelectedPeriod string
-	Periods        []PeriodOption
-
-	// Routing income
-	RoutingFeeSats int64
-	RoutingFeeUSD  float64
-	RoutedCount    int64
-
-	// Exchange summary
-	PurchasedSats        int64
-	ReceivedSats         int64
-	SoldSats             int64
-	SentSats             int64
-	TotalCostBasisUSD    float64
-	TotalSaleProceedsUSD float64
-	FeesPaidUSD          float64
-	StrikeFeesPaidUSD    float64
-	RiverFeesPaidUSD     float64
-	CoinbaseFeesPaidUSD  float64
-
-	// Net position
-	NetBTCSats int64
-	NetUSDSpent float64
-}
-
-// HandlePLPage serves GET /pl.
-func (h *Handler) HandlePLPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	now := time.Now()
-
-	// Parse period
-	period := r.URL.Query().Get("period")
-	var since time.Time
-	switch period {
-	case "90d":
-		since = now.AddDate(0, 0, -90)
-	case "ytd":
-		since = time.Date(now.Year(), 1, 1, 0, 0, 0, 0, time.UTC)
-	case "all":
-		since = time.Time{}
-	default:
-		period = "30d"
-		since = now.AddDate(0, 0, -30)
-	}
-
-	periods := []PeriodOption{
-		{Value: "30d", Label: "30 days", Active: period == "30d"},
-		{Value: "90d", Label: "90 days", Active: period == "90d"},
-		{Value: "ytd", Label: "YTD", Active: period == "ytd"},
-		{Value: "all", Label: "All time", Active: period == "all"},
-	}
-
-	data := PLPageData{
-		SelectedPeriod: period,
-		Periods:        periods,
-	}
-
-	// Concurrent data fetch
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	type result struct {
-		feesMsat       int64
-		routedCount    int64
-		strikeSummary  *db.ExchangeSummaryResult
-		riverSummary    *db.ExchangeSummaryResult
-		coinbaseSummary *db.ExchangeSummaryResult
-		nodeInfo        *lnd.NodeInfo
-		lastSynced     time.Time
-		btcPrice       float64
-		priceFetched   time.Time
-	}
-	var res result
-
-	fetch := func(fn func()) {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			fn()
-		}()
-	}
-
-	fetch(func() {
-		fees, routed, err := h.store.FeeSummary(ctx, since)
-		if err == nil {
-			mu.Lock()
-			res.feesMsat = fees
-			res.routedCount = routed
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		summary, err := h.store.ExchangeSummary(ctx, "strike", since)
-		if err == nil {
-			mu.Lock()
-			res.strikeSummary = summary
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		summary, err := h.store.ExchangeSummary(ctx, "river", since)
-		if err == nil {
-			mu.Lock()
-			res.riverSummary = summary
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		summary, err := h.store.ExchangeSummary(ctx, "coinbase", since)
-		if err == nil {
-			mu.Lock()
-			res.coinbaseSummary = summary
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		info := h.getNodeInfo(ctx)
-		if info != nil {
-			mu.Lock()
-			res.nodeInfo = info
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		t, err := h.store.LastSyncedAt(ctx)
-		if err == nil {
-			mu.Lock()
-			res.lastSynced = t
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		price, err := h.price.GetBTCPrice(ctx)
-		if err == nil {
-			mu.Lock()
-			res.btcPrice = price
-			res.priceFetched = h.price.FetchedAt()
-			mu.Unlock()
-		}
-	})
-
-	wg.Wait()
-
-	// Assemble template data
-	if res.nodeInfo != nil {
-		data.NodeAlias = res.nodeInfo.Alias
-		data.NodePubKey = res.nodeInfo.PubKey
-		data.NodeSynced = res.nodeInfo.Synced
-		data.BlockHeight = res.nodeInfo.BlockHeight
-		data.LNDVersion = res.nodeInfo.Version
-	}
-
-	data.LastSyncedAt = res.lastSynced
-	data.RoutingFeeSats = res.feesMsat / 1000
-	data.RoutedCount = res.routedCount
-
-	// Aggregate exchange summaries across all sources
-	for _, s := range []*db.ExchangeSummaryResult{res.strikeSummary, res.riverSummary, res.coinbaseSummary} {
-		if s != nil {
-			data.PurchasedSats += s.PurchasedSats
-			data.ReceivedSats += s.ReceivedSats
-			data.SoldSats += s.SoldSats
-			data.SentSats += s.SentSats
-			data.TotalCostBasisUSD += s.TotalCostBasisUSD
-			data.TotalSaleProceedsUSD += s.TotalSaleProceedsUSD
-			data.FeesPaidUSD += s.FeesPaidUSD
-		}
-	}
-	if res.strikeSummary != nil {
-		data.StrikeFeesPaidUSD = res.strikeSummary.FeesPaidUSD
-	}
-	if res.riverSummary != nil {
-		data.RiverFeesPaidUSD = res.riverSummary.FeesPaidUSD
-	}
-	if res.coinbaseSummary != nil {
-		data.CoinbaseFeesPaidUSD = res.coinbaseSummary.FeesPaidUSD
-	}
-
-	// Net BTC = routing fees + purchased + received - sold - sent
-	data.NetBTCSats = data.RoutingFeeSats + data.PurchasedSats + data.ReceivedSats - data.SoldSats - data.SentSats
-	// Net USD spent = cost basis - sale proceeds + fees paid
-	data.NetUSDSpent = data.TotalCostBasisUSD - data.TotalSaleProceedsUSD + data.FeesPaidUSD
-
-	if res.btcPrice > 0 {
-		data.BTCPriceUSD = res.btcPrice
-		data.PriceFetchedAt = res.priceFetched
-		data.RoutingFeeUSD = msatToUSD(res.feesMsat, res.btcPrice)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "pl_layout", data); err != nil {
-		h.logger.Printf("failed to render P&L page: %v", err)
-	}
-}
-
-// TaxPageData holds data for the /tax page.
-type TaxPageData struct {
-	NodeAlias      string
-	NodePubKey     string
-	NodeSynced     bool
-	BlockHeight    uint32
-	LNDVersion     string
-	LastSyncedAt   time.Time
-	BTCPriceUSD    float64
-	PriceFetchedAt time.Time
-	Tier           string
-
-	Method       string
-	SelectedYear int
-	Years        []int
-
-	// Summary fields
-	TotalProceeds     float64
-	TotalCostBasis    float64
-	TotalGainLoss     float64
-	ShortTermGainLoss float64
-	LongTermGainLoss  float64
-	ShortTermCount    int
-	LongTermCount     int
-	TotalDisposals    int
-	UnmatchedSat      int64
-}
-
-// HandleTaxPage serves GET /tax.
-func (h *Handler) HandleTaxPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	method := r.URL.Query().Get("method")
-	if method != "lifo" {
-		method = "fifo"
-	}
-
-	yearStr := r.URL.Query().Get("year")
-	selectedYear := 0
-	if yearStr != "" {
-		if y, err := strconv.Atoi(yearStr); err == nil {
-			selectedYear = y
-		}
-	}
-
-	data := TaxPageData{
-		Method:       method,
-		SelectedYear: selectedYear,
-	}
-
-	// Fill node/price info
-	if info := h.getNodeInfo(ctx); info != nil {
-		data.NodeAlias = info.Alias
-		data.NodePubKey = info.PubKey
-		data.NodeSynced = info.Synced
-		data.BlockHeight = info.BlockHeight
-		data.LNDVersion = info.Version
-	}
-	data.LastSyncedAt, _ = h.store.LastSyncedAt(ctx)
-	if price, err := h.price.GetBTCPrice(ctx); err == nil {
-		data.BTCPriceUSD = price
-		data.PriceFetchedAt = h.price.FetchedAt()
-	}
-	data.Tier = string(TierFromContext(ctx))
-
-	// Generate year options
-	now := time.Now()
-	for y := now.Year(); y >= now.Year()-5; y-- {
-		data.Years = append(data.Years, y)
-	}
-
-	// Compute tax summary if taxStore is available
-	if h.taxStore != nil {
-		dbLots, err := h.taxStore.ListBTCLots(ctx)
-		if err == nil {
-			dbDisposals, err := h.taxStore.ListDisposals(ctx)
-			if err == nil {
-				lots := make([]tax.Lot, len(dbLots))
-				for i, l := range dbLots {
-					lots[i] = tax.Lot{
-						ID: l.ID, AcquiredAt: l.AcquiredAt, AmountSat: l.AmountSat,
-						PriceUSD: l.PriceUSD, Source: l.Source, ExternalID: l.ExternalID,
-					}
-				}
-				disposals := make([]tax.Disposal, len(dbDisposals))
-				for i, d := range dbDisposals {
-					disposals[i] = tax.Disposal{
-						DisposedAt: d.DisposedAt, AmountSat: d.AmountSat, ProceedsUSD: d.ProceedsUSD,
-						TxType: d.TxType, Source: d.Source, ExternalID: d.ExternalID,
-					}
-				}
-
-				taxMethod := tax.FIFO
-				if method == "lifo" {
-					taxMethod = tax.LIFO
-				}
-
-				result, err := tax.Match(lots, disposals, taxMethod)
-				if err == nil {
-					// Filter by year if specified
-					if selectedYear > 0 {
-						var filtered []tax.TaxableEvent
-						for _, e := range result.Events {
-							if e.DisposedAt.Year() == selectedYear {
-								filtered = append(filtered, e)
-							}
-						}
-						result.Events = filtered
-					}
-
-					summary := tax.Summarize(result)
-					data.TotalProceeds = summary.TotalProceeds
-					data.TotalCostBasis = summary.TotalCostBasis
-					data.TotalGainLoss = summary.TotalGainLoss
-					data.ShortTermGainLoss = summary.ShortTermGainLoss
-					data.LongTermGainLoss = summary.LongTermGainLoss
-					data.ShortTermCount = summary.ShortTermCount
-					data.LongTermCount = summary.LongTermCount
-					data.TotalDisposals = summary.TotalDisposals
-					data.UnmatchedSat = summary.UnmatchedSat
-				}
-			}
-		}
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "tax_layout", data); err != nil {
-		h.logger.Printf("failed to render tax page: %v", err)
-	}
-}
-
-// WalletsPageData holds data for the /wallets page.
-type WalletsPageData struct {
-	NodeAlias      string
-	NodePubKey     string
-	NodeSynced     bool
-	BlockHeight    uint32
-	LNDVersion     string
-	LastSyncedAt   time.Time
-	BTCPriceUSD    float64
-	PriceFetchedAt time.Time
-
-	Wallets           []db.WatchedWallet
-	TotalBalanceSats  int64
-	TotalBalanceUSD   float64
-	ElectrumAvailable bool
-	Toast             string
-}
-
-// HandleWalletsPage serves GET /wallets.
-func (h *Handler) HandleWalletsPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	data := WalletsPageData{
-		ElectrumAvailable: h.walletScanner != nil,
-		Toast:             r.URL.Query().Get("toast"),
-	}
-
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	fetch := func(fn func()) {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			fn()
-		}()
-	}
-
-	if h.walletStore != nil {
-		fetch(func() {
-			wallets, err := h.walletStore.ListWallets(ctx)
-			if err == nil {
-				mu.Lock()
-				data.Wallets = wallets
-				mu.Unlock()
-			}
-		})
-
-		fetch(func() {
-			total, err := h.walletStore.TotalWatchedBalance(ctx)
-			if err == nil {
-				mu.Lock()
-				data.TotalBalanceSats = total
-				mu.Unlock()
-			}
-		})
-	}
-
-	fetch(func() {
-		info := h.getNodeInfo(ctx)
-		if info != nil {
-			mu.Lock()
-			data.NodeAlias = info.Alias
-			data.NodePubKey = info.PubKey
-			data.NodeSynced = info.Synced
-			data.BlockHeight = info.BlockHeight
-			data.LNDVersion = info.Version
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		t, err := h.store.LastSyncedAt(ctx)
-		if err == nil {
-			mu.Lock()
-			data.LastSyncedAt = t
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		price, err := h.price.GetBTCPrice(ctx)
-		if err == nil {
-			mu.Lock()
-			data.BTCPriceUSD = price
-			data.PriceFetchedAt = h.price.FetchedAt()
-			mu.Unlock()
-		}
-	})
-
-	wg.Wait()
-
-	if data.BTCPriceUSD > 0 {
-		data.TotalBalanceUSD = satsToUSD(data.TotalBalanceSats, data.BTCPriceUSD)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "wallets_layout", data); err != nil {
-		h.logger.Printf("failed to render wallets page: %v", err)
-	}
-}
-
-// HandleAddWallet handles POST /api/wallets.
-func (h *Handler) HandleAddWallet(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		h.writeError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-	if h.walletStore == nil {
-		h.writeError(w, http.StatusServiceUnavailable, "wallet tracking not available")
-		return
-	}
-
-	if err := r.ParseForm(); err != nil {
-		h.writeError(w, http.StatusBadRequest, "invalid form data")
-		return
-	}
-
-	label := r.FormValue("label")
-	value := r.FormValue("value")
-
-	if label == "" || value == "" {
-		h.writeError(w, http.StatusBadRequest, "label and value are required")
-		return
-	}
-
-	// Detect wallet type and derivation
-	walletType := "address"
-	derivationType := "bip84"
-
-	if isDescriptor(value) {
-		walletType = "descriptor"
-		derivationType = ""
-	} else if len(value) > 100 {
-		// Likely an extended public key
-		walletType = "xpub"
-		if len(value) >= 4 {
-			switch value[:4] {
-			case "zpub":
-				derivationType = "bip84"
-			case "ypub":
-				derivationType = "bip49"
-			case "xpub":
-				derivationType = "bip44"
-			}
-		}
-	}
-
-	id, err := h.walletStore.AddWallet(r.Context(), label, walletType, value, derivationType)
-	if err != nil {
-		h.logger.Printf("add wallet failed: %v", err)
-		h.writeError(w, http.StatusInternalServerError, "failed to add wallet")
-		return
-	}
-
-	// Trigger a background scan if scanner is available
-	if h.walletScanner != nil {
-		go h.scanWallet(id, label, walletType, value, derivationType)
-	}
-
-	// Redirect back immediately with a toast notification
-	if r.Header.Get("HX-Request") == "true" {
-		w.Header().Set("HX-Redirect", "/wallets?toast=scanning")
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-	http.Redirect(w, r, "/wallets", http.StatusSeeOther)
-}
-
-// HandleRemoveWallet handles POST /api/wallets/delete.
-func (h *Handler) HandleRemoveWallet(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		h.writeError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-	if h.walletStore == nil {
-		h.writeError(w, http.StatusServiceUnavailable, "wallet tracking not available")
-		return
-	}
-
-	if err := r.ParseForm(); err != nil {
-		h.writeError(w, http.StatusBadRequest, "invalid form data")
-		return
-	}
-
-	idStr := r.FormValue("id")
-	id, _ := strconv.ParseInt(idStr, 10, 64)
-	if id <= 0 {
-		h.writeError(w, http.StatusBadRequest, "invalid wallet id: "+idStr)
-		return
-	}
-
-	if err := h.walletStore.RemoveWallet(r.Context(), id); err != nil {
-		h.logger.Printf("remove wallet %d failed: %v", id, err)
-		h.writeError(w, http.StatusInternalServerError, "failed to remove wallet")
-		return
-	}
-
-	if r.Header.Get("HX-Request") == "true" {
-		w.Header().Set("HX-Redirect", "/wallets")
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-	http.Redirect(w, r, "/wallets", http.StatusSeeOther)
-}
-
-// HandleRefreshWallet handles POST /api/wallets/refresh.
-func (h *Handler) HandleRefreshWallet(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		h.writeError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-	if h.walletStore == nil || h.walletScanner == nil {
-		h.writeError(w, http.StatusServiceUnavailable, "wallet scanning not available")
-		return
-	}
-
-	if err := r.ParseForm(); err != nil {
-		h.writeError(w, http.StatusBadRequest, "invalid form data")
-		return
-	}
-
-	idStr := r.FormValue("id")
-	id, _ := strconv.ParseInt(idStr, 10, 64)
-	if id <= 0 {
-		h.writeError(w, http.StatusBadRequest, "invalid wallet id")
-		return
-	}
-
-	wallet, err := h.walletStore.GetWallet(r.Context(), id)
-	if err != nil || wallet == nil {
-		h.writeError(w, http.StatusNotFound, "wallet not found")
-		return
-	}
-
-	go h.scanWallet(wallet.ID, wallet.Label, wallet.Type, wallet.Value, wallet.DerivationType)
-
-	if r.Header.Get("HX-Request") == "true" {
-		w.Header().Set("HX-Redirect", "/wallets?toast=scanning")
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-	http.Redirect(w, r, "/wallets?toast=scanning", http.StatusSeeOther)
-}
-
-// HandleRefreshAll handles POST /api/wallets/refresh-all.
-// Kicks off background scans for all wallets and redirects immediately.
-func (h *Handler) HandleRefreshAll(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		h.writeError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-	if h.walletStore == nil || h.walletScanner == nil {
-		h.writeError(w, http.StatusServiceUnavailable, "wallet scanning not available")
-		return
-	}
-
-	wallets, err := h.walletStore.ListWallets(r.Context())
-	if err != nil {
-		h.logger.Printf("refresh-all: list wallets failed: %v", err)
-		h.writeError(w, http.StatusInternalServerError, "failed to list wallets")
-		return
-	}
-
-	// Scan all wallets sequentially in the background
-	go func() {
-		for _, wlt := range wallets {
-			h.scanWallet(wlt.ID, wlt.Label, wlt.Type, wlt.Value, wlt.DerivationType)
-		}
-		h.logger.Printf("refresh-all: completed scanning %d wallets", len(wallets))
-	}()
-
-	h.logger.Printf("refresh-all: started background scan for %d wallets", len(wallets))
-
-	if r.Header.Get("HX-Request") == "true" {
-		w.Header().Set("HX-Redirect", "/wallets?toast=scanning")
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-	http.Redirect(w, r, "/wallets?toast=scanning", http.StatusSeeOther)
-}
-
-// scanWallet runs a balance scan for a single wallet in the background.
-// It uses context.Background() so it survives after the HTTP request ends.
-func (h *Handler) scanWallet(id int64, label, walletType, value, derivationType string) {
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-	defer cancel()
-
-	var balance int64
-	var err error
-
-	switch walletType {
-	case "address":
-		balance, err = h.walletScanner.ScanAddress(ctx, value)
-	case "xpub":
-		balance, err = h.walletScanner.ScanXpub(ctx, value, derivationType)
-	case "descriptor":
-		balance, err = h.walletScanner.ScanDescriptor(ctx, value)
-	default:
-		h.logger.Printf("scan %q: unknown wallet type %q", label, walletType)
-		return
-	}
-
-	if err != nil {
-		h.logger.Printf("scan %q: %v", label, err)
-		return
-	}
-
-	if err := h.walletStore.UpdateWalletBalance(context.Background(), id, balance); err != nil {
-		h.logger.Printf("scan %q: failed to save balance: %v", label, err)
-		return
-	}
-
-	h.logger.Printf("scan %q: %d sats", label, balance)
-}
-
-// ExchangeDetailData holds data for the exchange transaction detail page.
-type ExchangeDetailData struct {
-	Source      string
-	SourceLabel string
-
-	// Summary stats
-	Summary    *db.ExchangeSummaryResult
-	BalanceSats int64
-
-	// Transaction list
-	Transactions []db.ExchangeTransaction
-	Total        int
-	Page         int
-	Limit        int
-	TotalPages   int
-
-	// Price
-	BTCPriceUSD    float64
-	PriceFetchedAt time.Time
-}
-
-// HandleExchangeDetail serves GET /exchange/{source}.
-func (h *Handler) HandleExchangeDetail(w http.ResponseWriter, r *http.Request) {
-	source := strings.TrimPrefix(r.URL.Path, "/exchange/")
-	if source == "" || (source != "strike" && source != "river" && source != "coinbase" && source != "swan") {
-		http.NotFound(w, r)
-		return
-	}
-
-	labels := map[string]string{"strike": "Strike", "river": "River", "coinbase": "Coinbase", "swan": "Swan"}
-	ctx := r.Context()
-
-	page, _ := strconv.Atoi(r.URL.Query().Get("page"))
-	if page < 1 {
-		page = 1
-	}
-	limit := 50
-	offset := (page - 1) * limit
-
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-	data := ExchangeDetailData{
-		Source:      source,
-		SourceLabel: labels[source],
-		Page:        page,
-		Limit:       limit,
-	}
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		summary, err := h.store.ExchangeSummary(ctx, source, time.Time{})
-		if err == nil {
-			mu.Lock()
-			data.Summary = summary
-			mu.Unlock()
-		}
-	}()
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		bal, err := h.store.ExchangeBalance(ctx, source)
-		if err == nil {
-			mu.Lock()
-			data.BalanceSats = bal
-			mu.Unlock()
-		}
-	}()
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		result, err := h.store.ListExchangeTransactions(ctx, source, limit, offset)
-		if err == nil {
-			mu.Lock()
-			data.Transactions = result.Transactions
-			data.Total = result.Total
-			mu.Unlock()
-		}
-	}()
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		price, err := h.price.GetBTCPrice(ctx)
-		if err == nil {
-			mu.Lock()
-			data.BTCPriceUSD = price
-			data.PriceFetchedAt = h.price.FetchedAt()
-			mu.Unlock()
-		}
-	}()
-
-	wg.Wait()
-
-	if data.Total > 0 {
-		data.TotalPages = (data.Total + limit - 1) / limit
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "exchange_detail_layout", data); err != nil {
-		h.logger.Printf("failed to render exchange detail: %v", err)
-	}
-}
-
-// WalletDetailData holds data for the wallet detail page.
-type WalletDetailData struct {
-	Wallet      *db.WatchedWallet
-	BTCPriceUSD float64
-	BalanceUSD  float64
-}
-
-// HandleWalletDetail serves GET /wallets/{id}.
-func (h *Handler) HandleWalletDetail(w http.ResponseWriter, r *http.Request) {
-	idStr := strings.TrimPrefix(r.URL.Path, "/wallets/")
-	id, err := strconv.ParseInt(idStr, 10, 64)
-	if err != nil || id < 1 {
-		http.NotFound(w, r)
-		return
-	}
-
-	if h.walletStore == nil {
-		http.NotFound(w, r)
-		return
-	}
-
-	ctx := r.Context()
-	wallet, err := h.walletStore.GetWallet(ctx, id)
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-
-	data := WalletDetailData{Wallet: wallet}
-	price, err := h.price.GetBTCPrice(ctx)
-	if err == nil && price > 0 {
-		data.BTCPriceUSD = price
-		data.BalanceUSD = float64(wallet.BalanceSats) / 1e8 * price
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "wallet_detail_layout", data); err != nil {
-		h.logger.Printf("failed to render wallet detail: %v", err)
-	}
-}
-
-// isDescriptor returns true if the value looks like a Bitcoin output descriptor.
-func isDescriptor(v string) bool {
-	prefixes := []string{
-		"wpkh(", "sh(", "wsh(", "pk(", "pkh(", "combo(",
-		"multi(", "sortedmulti(", "tr(", "addr(", "raw(",
-	}
-	for _, p := range prefixes {
-		if strings.HasPrefix(v, p) {
-			return true
-		}
-	}
-	return false
-}
-
-// parseDateParam parses a "YYYY-MM-DD" query parameter.
-func parseDateParam(r *http.Request, key string, defaultVal time.Time) (time.Time, error) {
-	v := r.URL.Query().Get(key)
-	if v == "" {
-		return defaultVal, nil
-	}
-	return time.Parse("2006-01-02", v)
-}
-
-// SettingsPageData holds data for the settings page.
-type SettingsPageData struct {
-	Tier             string
-	LicenseKey       string // masked
-	CheckoutURL      string // Stripe checkout URL for subscribing
-	MonarchToken     string
-	MonarchConnected bool
-	MonarchUnlocked  bool
-	Message          string
-	Error            string
-	// Auto-activation from Stripe redirect
-	AutoActivateKey  string
-	// Monarch transaction sync
-	MonarchSyncTypes    []string // currently selected tx types
-	MonarchSyncSources  []string // currently selected sources
-	AvailableTxTypes    []string // all tx types from data
-	AvailableSources    []string // all sources from data
-	MonarchSyncedCount  int      // how many transactions have been synced
-}
-
-// HandleSettingsPage serves GET /settings.
-func (h *Handler) HandleSettingsPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	tier := TierFromContext(ctx)
-
-	// Handle auto-activation from Stripe redirect (?license_key=sk_xxx)
-	autoKey := r.URL.Query().Get("license_key")
-	if autoKey != "" && h.settingsStore != nil && h.licenseChecker != nil {
-		_ = h.settingsStore.SetSetting(ctx, "license_key", autoKey)
-		if err := h.licenseChecker.SetKeyAndVerify(ctx, autoKey); err != nil {
-			h.logger.Printf("auto-activate license: %v", err)
-		} else {
-			tier = h.licenseChecker.CurrentTier()
-		}
-	}
-
-	data := SettingsPageData{
-		Tier:            string(tier),
-		MonarchUnlocked: license.TierAtLeast(tier, license.TierPower),
-		AutoActivateKey: autoKey,
-	}
-
-	// Show masked license key
-	if h.licenseChecker != nil && h.licenseChecker.LicenseKey() != "" {
-		data.LicenseKey = maskToken(h.licenseChecker.LicenseKey())
-	}
-
-	if data.MonarchUnlocked && h.settingsStore != nil {
-		token, _ := h.settingsStore.GetSetting(r.Context(), "monarch_token")
-		if token != "" {
-			data.MonarchToken = maskToken(token)
-			data.MonarchConnected = true
-		}
-
-		// Load tx sync preferences
-		syncTypes, _ := h.settingsStore.GetSetting(r.Context(), "monarch_sync_types")
-		if syncTypes != "" {
-			data.MonarchSyncTypes = strings.Split(syncTypes, ",")
-		}
-		syncSources, _ := h.settingsStore.GetSetting(r.Context(), "monarch_sync_sources")
-		if syncSources != "" {
-			data.MonarchSyncSources = strings.Split(syncSources, ",")
-		}
-
-		// Load available tx types and sources from data
-		if h.monarchTxStore != nil {
-			sources, txTypes, _ := h.monarchTxStore.DistinctTransactionValues(r.Context())
-			data.AvailableTxTypes = txTypes
-			data.AvailableSources = sources
-			count, _ := h.monarchTxStore.MonarchSyncedCount(r.Context())
-			data.MonarchSyncedCount = count
-		}
-	}
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "settings_layout", data); err != nil {
-		h.logger.Printf("failed to render settings page: %v", err)
-	}
-}
-
-// HandleMonarchSave handles POST /api/monarch/save — logs into Monarch and stores the auth token.
-// Supports two-step flow: if OTP is required, returns an OTP input form.
-func (h *Handler) HandleMonarchSave(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	if h.settingsStore == nil {
-		http.Error(w, "settings not available", http.StatusInternalServerError)
-		return
-	}
-
-	email := strings.TrimSpace(r.FormValue("monarch_email"))
-	password := r.FormValue("monarch_password")
-	otpCode := strings.TrimSpace(r.FormValue("monarch_otp"))
-
-	if email == "" || password == "" {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Email and password are required.</div>`)
-		return
-	}
-
-	var syncer *monarch.Syncer
-	var token string
-	var err error
-
-	if otpCode != "" && h.pendingMonarch != nil {
-		// Step 2: complete login with OTP using the same client session
-		syncer, token, err = h.pendingMonarch.CompleteOTP(r.Context(), email, password, otpCode)
-		h.pendingMonarch = nil
-	} else {
-		// Step 1: attempt login
-		var pending *monarch.PendingClient
-		syncer, token, pending, err = monarch.NewSyncerWithLogin(r.Context(), email, password, "")
-		if errors.Is(err, monarch.ErrOTPRequired) {
-			// Store pending client for step 2
-			h.pendingMonarch = pending
-			w.Header().Set("Content-Type", "text/html; charset=utf-8")
-			fmt.Fprintf(w, `<div class="alert alert-info">Check your email for a verification code.</div>
-<form hx-post="/api/monarch/save" hx-target="#monarch-result" style="display: flex; gap: 8px; align-items: flex-end; margin-top: 8px;">
-  <input type="hidden" name="monarch_email" value="%s">
-  <input type="hidden" name="monarch_password" value="%s">
-  <div>
-    <label style="font-size: 0.8rem; color: #aaa;">Verification Code</label>
-    <input type="text" name="monarch_otp" placeholder="6-digit code" style="width: 160px; margin-top: 4px;" required autofocus>
-  </div>
-  <button type="submit" class="btn btn-primary">Verify</button>
-</form>`, email, password)
-			return
-		}
-	}
-
-	if err != nil {
-		h.logger.Printf("monarch login failed: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprintf(w, `<div class="alert alert-error">Login failed: %s</div>`, err.Error())
-		return
-	}
-
-	// Store the token
-	if err := h.settingsStore.SetSetting(r.Context(), "monarch_token", token); err != nil {
-		h.logger.Printf("failed to save monarch token: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Login succeeded but failed to save token.</div>`)
-		return
-	}
-
-	// Hot-swap the syncer (no restart needed)
-	h.monarchSyncer = syncer
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	fmt.Fprint(w, monarchConnectedHTML)
-}
-
-const monarchConnectedHTML = `<div class="alert alert-success">Connected to Monarch Money.</div>
-<div style="display: flex; gap: 8px; margin-top: 12px;">
-  <button hx-post="/api/monarch/sync" hx-target="#monarch-result" class="btn btn-primary">Sync Now</button>
-  <button hx-post="/api/monarch/disconnect" hx-target="#monarch-result" hx-confirm="Disconnect Monarch Money?" class="btn btn-danger">Disconnect</button>
-</div>`
-
-// HandleMonarchToken handles POST /api/monarch/token — saves a manually-provided token.
-func (h *Handler) HandleMonarchToken(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	if h.settingsStore == nil {
-		http.Error(w, "settings not available", http.StatusInternalServerError)
-		return
-	}
-
-	token := strings.TrimSpace(r.FormValue("monarch_token"))
-	if token == "" {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Token is required.</div>`)
-		return
-	}
-
-	// Create syncer from token to verify it works
-	syncer, err := monarch.NewSyncer(token, "")
-	if err != nil {
-		h.logger.Printf("monarch token invalid: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprintf(w, `<div class="alert alert-error">Invalid token: %s</div>`, err.Error())
-		return
-	}
-
-	if err := h.settingsStore.SetSetting(r.Context(), "monarch_token", token); err != nil {
-		h.logger.Printf("failed to save monarch token: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Failed to save token.</div>`)
-		return
-	}
-
-	h.monarchSyncer = syncer
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	fmt.Fprint(w, monarchConnectedHTML)
-}
-
-// HandleMonarchDisconnect handles POST /api/monarch/disconnect — removes the token.
-func (h *Handler) HandleMonarchDisconnect(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	if h.settingsStore == nil {
-		http.Error(w, "settings not available", http.StatusInternalServerError)
-		return
-	}
-
-	if err := h.settingsStore.SetSetting(r.Context(), "monarch_token", ""); err != nil {
-		h.logger.Printf("failed to clear monarch token: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Failed to disconnect.</div>`)
-		return
-	}
-
-	// Clear syncer so the background loop stops syncing too
-	h.SetMonarchSyncer(nil)
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	fmt.Fprint(w, `<div class="alert alert-success">Monarch disconnected.</div>`)
-}
-
-// HandleMonarchSync handles POST /api/monarch/sync — triggers a manual sync.
-func (h *Handler) HandleMonarchSync(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	// Create syncer on-demand from stored token if not already set
-	if h.monarchSyncer == nil && h.settingsStore != nil {
-		token, _ := h.settingsStore.GetSetting(r.Context(), "monarch_token")
-		if token != "" {
-			syncer, err := monarch.NewSyncer(token, "")
-			if err == nil {
-				h.monarchSyncer = syncer
-			}
-		}
-	}
-
-	if h.monarchSyncer == nil {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Monarch not connected. Log in first.</div>`)
-		return
-	}
-
-	// Get total BTC from portfolio
-	totalSats, err := h.getTotalSats(r.Context())
-	if err != nil {
-		h.logger.Printf("monarch sync: failed to get total sats: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div class="alert alert-error">Failed to calculate total BTC.</div>`)
-		return
-	}
-
-	totalBTC := float64(totalSats) / 1e8
-
-	if err := h.monarchSyncer.SyncHolding(r.Context(), totalBTC); err != nil {
-		h.logger.Printf("monarch sync failed: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprintf(w, `<div class="alert alert-error">Sync failed: %s</div>`, err.Error())
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	fmt.Fprintf(w, `<div class="alert alert-success">Synced %.8f BTC to Monarch Money as a holding.</div>`, totalBTC)
-}
-
-// getTotalSats returns the total sats across all sources.
-func (h *Handler) getTotalSats(ctx context.Context) (int64, error) {
-	var total int64
-
-	// On-chain wallet balance
-	if bal, err := h.store.LatestWalletBalance(ctx); err == nil && bal != nil {
-		total += bal.TotalSat
-	}
-
-	// Exchange balances
-	for _, src := range []string{"strike", "river", "coinbase", "swan"} {
-		if bal, err := h.store.ExchangeBalance(ctx, src); err == nil {
-			total += bal
-		}
-	}
-
-	// Cold storage (watched wallets)
-	if h.walletStore != nil {
-		if bal, err := h.walletStore.TotalWatchedBalance(ctx); err == nil {
-			total += bal
-		}
-	}
-
-	return total, nil
-}
-
-// HandleMonarchSyncTypes handles POST /api/monarch/sync-types — saves which tx types to sync.
-func (h *Handler) HandleMonarchSyncTypes(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	if h.settingsStore == nil {
-		http.Error(w, "settings not available", http.StatusInternalServerError)
-		return
-	}
-
-	if err := r.ParseForm(); err != nil {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">Failed to parse form.</div>`)
-		return
-	}
-
-	selectedTypes := r.Form["sync_types"]
-	selectedSources := r.Form["sync_sources"]
-
-	if err := h.settingsStore.SetSetting(r.Context(), "monarch_sync_types", strings.Join(selectedTypes, ",")); err != nil {
-		h.logger.Printf("monarch: failed to save sync types: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">Failed to save preferences.</div>`)
-		return
-	}
-	if err := h.settingsStore.SetSetting(r.Context(), "monarch_sync_sources", strings.Join(selectedSources, ",")); err != nil {
-		h.logger.Printf("monarch: failed to save sync sources: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">Failed to save preferences.</div>`)
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if len(selectedTypes) == 0 && len(selectedSources) == 0 {
-		fmt.Fprint(w, `<div style="color:var(--green);">Transaction sync disabled — no filters selected.</div>`)
-	} else {
-		fmt.Fprintf(w, `<div style="color:var(--green);">Saved preferences: %d type(s), %d source(s).</div>`, len(selectedTypes), len(selectedSources))
-	}
-}
-
-// HandleMonarchTxSync handles POST /api/monarch/tx-sync — syncs transactions to Monarch.
-func (h *Handler) HandleMonarchTxSync(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	// Ensure Monarch syncer is available
-	if h.monarchSyncer == nil && h.settingsStore != nil {
-		token, _ := h.settingsStore.GetSetting(r.Context(), "monarch_token")
-		if token != "" {
-			syncer, err := monarch.NewSyncer(token, "")
-			if err == nil {
-				h.monarchSyncer = syncer
-			}
-		}
-	}
-	if h.monarchSyncer == nil {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">Monarch not connected. Log in first.</div>`)
-		return
-	}
-
-	if h.settingsStore == nil || h.monarchTxStore == nil {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">Transaction sync not available.</div>`)
-		return
-	}
-
-	// Read selected sync types and sources
-	syncTypes, _ := h.settingsStore.GetSetting(r.Context(), "monarch_sync_types")
-	if syncTypes == "" {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">No transaction types selected. Save preferences first.</div>`)
-		return
-	}
-
-	types := strings.Split(syncTypes, ",")
-	var sources []string
-	if v, _ := h.settingsStore.GetSetting(r.Context(), "monarch_sync_sources"); v != "" {
-		sources = strings.Split(v, ",")
-	}
-
-	// Get unsynced transactions
-	txns, err := h.monarchTxStore.ListUnsyncedTransactions(r.Context(), types, sources)
-	if err != nil {
-		h.logger.Printf("monarch tx sync: list unsynced: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--red);">Failed to query transactions.</div>`)
-		return
-	}
-
-	if len(txns) == 0 {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		fmt.Fprint(w, `<div style="color:var(--green);">All transactions are already synced.</div>`)
-		return
-	}
-
-	// Convert to TxToSync
-	toSync := make([]monarch.TxToSync, len(txns))
-	for i, tx := range txns {
-		toSync[i] = monarch.TxToSync{
-			SourceID:  tx.SourceID,
-			Source:    tx.Source,
-			TxType:   tx.TxType,
-			Time:     tx.Time,
-			AmountUSD: tx.AmountUSD,
-			Memo:     tx.Memo,
-		}
-	}
-
-	// Respond immediately — sync runs in the background
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	fmt.Fprintf(w, `<div style="color:var(--accent);">Syncing %d transactions to Monarch in the background. You can leave this page.</div>`, len(toSync))
-
-	// Capture references for the goroutine
-	syncer := h.monarchSyncer
-	txStore := h.monarchTxStore
-	logger := h.logger
-
-	go func() {
-		ctx := context.Background()
-		result, synced, err := syncer.SyncTransactions(ctx, toSync)
-		if err != nil {
-			logger.Printf("monarch tx sync failed: %v", err)
-			return
-		}
-
-		for sourceID, monarchTxID := range synced {
-			if err := txStore.MarkTransactionSynced(ctx, sourceID, monarchTxID); err != nil {
-				logger.Printf("monarch: failed to mark %s synced: %v", sourceID, err)
-			}
-		}
-
-		logger.Printf("monarch tx sync complete: %d created, %d skipped, %d errors", result.Created, result.Skipped, result.Errors)
-	}()
-}
-
-// TransactionsPageData holds data for the /transactions page.
-type TransactionsPageData struct {
-	Transactions []db.UnifiedTransaction
-	Total        int
-	Page         int
-	Limit        int
-	TotalPages   int
-	BTCPriceUSD  float64
-
-	// Current filter/sort state (for preserving in links and form)
-	DateFrom string
-	DateTo   string
-	TxType   string
-	Source   string
-	Search   string
-	Flow     string // "inflow", "outflow", or ""
-	SortCol  string
-	SortDir  string
-
-	// Available filter options
-	Sources []string
-	TxTypes []string
-}
-
-// HandleTransactionsPage serves GET /transactions.
-func (h *Handler) HandleTransactionsPage(w http.ResponseWriter, r *http.Request) {
-	if h.txStore == nil {
-		http.Error(w, "transaction store not available", http.StatusInternalServerError)
-		return
-	}
-
-	ctx := r.Context()
-	page := parseIntParam(r, "page", 1)
-	if page < 1 {
-		page = 1
-	}
-	limit := 50
-	offset := (page - 1) * limit
-
-	q := r.URL.Query()
-	sortCol := q.Get("sort")
-	sortDir := q.Get("dir")
-	if sortDir == "" {
-		sortDir = "desc"
-	}
-
-	flow := q.Get("flow")
-	f := db.TransactionFilter{
-		DateFrom: q.Get("from"),
-		DateTo:   q.Get("to"),
-		TxType:   q.Get("type"),
-		Source:   q.Get("source"),
-		Search:   q.Get("q"),
-		Flow:     flow,
-		SortCol:  sortCol,
-		SortDir:  sortDir,
-		Limit:    limit,
-		Offset:   offset,
-	}
-
-	result, err := h.txStore.ListUnifiedTransactions(ctx, f)
-	if err != nil {
-		h.logger.Printf("transactions page: %v", err)
-		http.Error(w, "failed to load transactions", http.StatusInternalServerError)
-		return
-	}
-
-	// Pull actual sources and types from the data
-	sources, txTypes, _ := h.txStore.DistinctTransactionValues(ctx)
-
-	data := TransactionsPageData{
-		Transactions: result.Transactions,
-		Total:        result.Total,
-		Page:         page,
-		Limit:        limit,
-		DateFrom:     f.DateFrom,
-		DateTo:       f.DateTo,
-		TxType:       f.TxType,
-		Source:       f.Source,
-		Search:       f.Search,
-		Flow:         flow,
-		SortCol:      sortCol,
-		SortDir:      sortDir,
-		Sources:      sources,
-		TxTypes:      txTypes,
-	}
-	if result.Total > 0 {
-		data.TotalPages = (result.Total + limit - 1) / limit
-	}
-
-	if price, err := h.price.GetBTCPrice(ctx); err == nil {
-		data.BTCPriceUSD = price
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "transactions_layout", data); err != nil {
-		h.logger.Printf("failed to render transactions page: %v", err)
-	}
-}
-
-// txNoteData is passed to the transaction_note_display and _edit partials.
-type txNoteData struct {
-	SourceID string
-	Note     string
-}
-
-// HandleTransactionNoteEdit serves GET /api/transactions/note/edit — returns edit form partial.
-func (h *Handler) HandleTransactionNoteEdit(w http.ResponseWriter, r *http.Request) {
-	sourceID := r.URL.Query().Get("source_id")
-	note := r.URL.Query().Get("note")
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "transaction_note_edit", txNoteData{SourceID: sourceID, Note: note}); err != nil {
-		h.logger.Printf("render note edit: %v", err)
-	}
-}
-
-// HandleTransactionNoteSave serves POST /api/transactions/note — saves and returns display partial.
-func (h *Handler) HandleTransactionNoteSave(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		h.writeError(w, http.StatusMethodNotAllowed, "method not allowed")
-		return
-	}
-	if h.txStore == nil {
-		h.writeError(w, http.StatusInternalServerError, "transaction store not available")
-		return
-	}
-
-	if err := r.ParseForm(); err != nil {
-		h.writeError(w, http.StatusBadRequest, "invalid form")
-		return
-	}
-
-	sourceID := r.FormValue("source_id")
-	note := r.FormValue("note")
-	if sourceID == "" {
-		h.writeError(w, http.StatusBadRequest, "source_id required")
-		return
-	}
-
-	if err := h.txStore.SetTransactionNote(r.Context(), sourceID, note); err != nil {
-		h.logger.Printf("set transaction note: %v", err)
-		h.writeError(w, http.StatusInternalServerError, "failed to save note")
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "transaction_note_display", txNoteData{SourceID: sourceID, Note: strings.TrimSpace(note)}); err != nil {
-		h.logger.Printf("render note display: %v", err)
-	}
-}
-
-// HandleTransferToggle serves POST /api/transactions/transfer — toggles is_transfer flag.
-func (h *Handler) HandleTransferToggle(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	sourceID := r.FormValue("source_id")
-	if sourceID == "" {
-		http.Error(w, "source_id required", http.StatusBadRequest)
-		return
-	}
-
-	// Toggle: read current, flip it
-	current, _ := h.store.GetTransferFlag(r.Context(), sourceID)
-	newVal := !current
-	if err := h.store.SetTransferFlag(r.Context(), sourceID, newVal); err != nil {
-		h.logger.Printf("set transfer flag: %v", err)
-		http.Error(w, "failed to update", http.StatusInternalServerError)
-		return
-	}
-
-	// Return updated transfer badge partial
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "transfer_badge", transferBadgeData{SourceID: sourceID, IsTransfer: newVal}); err != nil {
-		h.logger.Printf("render transfer badge: %v", err)
-	}
-}
-
-// HandleTransferBulk serves POST /api/transactions/transfer/bulk — marks multiple transactions.
-func (h *Handler) HandleTransferBulk(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	ctx := r.Context()
-	sourceID := r.FormValue("source_id")
-	candidateID := r.FormValue("candidate_id")
-	if sourceID == "" {
-		http.Error(w, "source_id required", http.StatusBadRequest)
-		return
-	}
-
-	if err := h.store.SetTransferFlag(ctx, sourceID, true); err != nil {
-		h.logger.Printf("bulk transfer: %v", err)
-		http.Error(w, "failed to update", http.StatusInternalServerError)
-		return
-	}
-	if candidateID != "" {
-		if err := h.store.SetTransferFlag(ctx, candidateID, true); err != nil {
-			h.logger.Printf("bulk transfer candidate: %v", err)
-		}
-	}
-
-	// Return updated badge for the primary transaction
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "transfer_badge", transferBadgeData{SourceID: sourceID, IsTransfer: true}); err != nil {
-		h.logger.Printf("render transfer badge: %v", err)
-	}
-	// OOB swap: also update the candidate's badge in-place (no refresh needed)
-	if candidateID != "" {
-		if err := h.renderer.Render(w, "transfer_badge_oob", transferBadgeData{SourceID: candidateID, IsTransfer: true}); err != nil {
-			h.logger.Printf("render transfer badge oob: %v", err)
-		}
-	}
-}
-
-// HandleTransferCandidates serves GET /api/transactions/transfer/candidates — auto-suggest.
-func (h *Handler) HandleTransferCandidates(w http.ResponseWriter, r *http.Request) {
-	sourceID := r.URL.Query().Get("source_id")
-	amountSat, _ := strconv.ParseInt(r.URL.Query().Get("amount_sat"), 10, 64)
-	tsStr := r.URL.Query().Get("ts")
-	ts, _ := time.Parse(time.RFC3339, tsStr)
-
-	if sourceID == "" || amountSat == 0 || ts.IsZero() {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		// Still render the popover so user sees "no matches" feedback
-		_ = h.renderer.Render(w, "transfer_candidates", transferCandidatesData{SourceID: sourceID})
-		return
-	}
-
-	candidates, err := h.store.ListTransferCandidates(r.Context(), sourceID, amountSat, ts)
-	if err != nil {
-		h.logger.Printf("transfer candidates: %v", err)
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "transfer_candidates", transferCandidatesData{
-		SourceID:   sourceID,
-		Candidates: candidates,
-	}); err != nil {
-		h.logger.Printf("render transfer candidates: %v", err)
-	}
-}
-
-type transferBadgeData struct {
-	SourceID   string
-	IsTransfer bool
-	AmountSat  int64
-	Time       time.Time
-}
-
-type transferCandidatesData struct {
-	SourceID   string
-	Candidates []db.TransferCandidate
-}
-
-// BreakdownItem represents a single source in the portfolio donut chart / table.
-type BreakdownItem struct {
-	Label     string
-	SourceKey string // machine key for API calls (e.g. "strike", "onchain")
-	Clickable bool   // false for channels/cold_storage (no transaction flows)
-	Sats      int64
-	USD       float64
-	Pct       float64
-}
-
-// PortfolioPageData holds data for the /portfolio page.
-type PortfolioPageData struct {
-	// Layout fields
-	Tier           string
-	NodeAlias      string
-	NodePubKey     string
-	NodeSynced     bool
-	BlockHeight    uint32
-	LNDVersion     string
-	BTCPriceUSD    float64
-	PriceFetchedAt time.Time
-	LastSyncedAt   time.Time
-
-	// Breakdown
-	Items       []BreakdownItem
-	DonutColors []string
-	TotalSats   int64
-	TotalUSD    float64
-
-	// Date filter for linking to transactions
-	DateFrom string
-
-	// Net flows
-	NetFlow *db.NetFlowResult
-
-	// Cost basis
-	TotalCostBasisUSD  float64
-	AvgCostPerBTC      float64
-	UnrealizedGainUSD  float64
-	CurrentValueUSD    float64
-
-	// Period selector
-	SelectedPeriod string
-	Periods        []PeriodOption
-}
-
-// HandlePortfolioPage serves GET /portfolio.
-func (h *Handler) HandlePortfolioPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	now := time.Now()
-
-	// Parse period
-	period := r.URL.Query().Get("period")
-	var since time.Time
-	switch period {
-	case "90d":
-		since = now.AddDate(0, 0, -90)
-	case "ytd":
-		since = time.Date(now.Year(), 1, 1, 0, 0, 0, 0, time.UTC)
-	case "all":
-		since = time.Time{}
-	default:
-		period = "30d"
-		since = now.AddDate(0, 0, -30)
-	}
-
-	dateFrom := ""
-	if !since.IsZero() {
-		dateFrom = since.Format("2006-01-02")
-	}
-
-	data := PortfolioPageData{
-		Tier:           string(TierFromContext(ctx)),
-		SelectedPeriod: period,
-		DateFrom:       dateFrom,
-		Periods: []PeriodOption{
-			{Value: "30d", Label: "30 days", Active: period == "30d"},
-			{Value: "90d", Label: "90 days", Active: period == "90d"},
-			{Value: "ytd", Label: "YTD", Active: period == "ytd"},
-			{Value: "all", Label: "All time", Active: period == "all"},
-		},
-	}
-
-	// Concurrent data fetch
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	fetch := func(fn func()) {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			fn()
-		}()
-	}
-
-	var breakdown *db.PortfolioBreakdown
-	fetch(func() {
-		b, err := h.store.PortfolioBreakdownQuery(ctx)
-		if err == nil {
-			mu.Lock()
-			breakdown = b
-			mu.Unlock()
-		}
-	})
-
-	var netFlow *db.NetFlowResult
-	fetch(func() {
-		nf, err := h.store.NetFlowSummary(ctx, since, true)
-		if err == nil {
-			mu.Lock()
-			netFlow = nf
-			mu.Unlock()
-		}
-	})
-
-	var position *db.PortfolioPositionResult
-	fetch(func() {
-		pos, err := h.store.PortfolioPosition(ctx, time.Time{})
-		if err == nil {
-			mu.Lock()
-			position = pos
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		info := h.getNodeInfo(ctx)
-		if info != nil {
-			mu.Lock()
-			data.NodeAlias = info.Alias
-			data.NodePubKey = info.PubKey
-			data.NodeSynced = info.Synced
-			data.BlockHeight = info.BlockHeight
-			data.LNDVersion = info.Version
-			mu.Unlock()
-		}
-	})
-
-	fetch(func() {
-		t, _ := h.store.LastSyncedAt(ctx)
-		mu.Lock()
-		data.LastSyncedAt = t
-		mu.Unlock()
-	})
-
-	fetch(func() {
-		price, err := h.price.GetBTCPrice(ctx)
-		if err == nil {
-			mu.Lock()
-			data.BTCPriceUSD = price
-			data.PriceFetchedAt = h.price.FetchedAt()
-			mu.Unlock()
-		}
-	})
-
-	wg.Wait()
-
-	// Build breakdown items
-	if breakdown != nil {
-		satsToUSD := func(sats int64) float64 {
-			if data.BTCPriceUSD <= 0 {
-				return 0
-			}
-			return float64(sats) / 1e8 * data.BTCPriceUSD
-		}
-
-		data.TotalSats = breakdown.TotalSats
-		data.TotalUSD = satsToUSD(breakdown.TotalSats)
-
-		type srcEntry struct {
-			label     string
-			sourceKey string
-			clickable bool
-			sats      int64
-		}
-		var entries []srcEntry
-		if breakdown.OnChainSats != 0 {
-			entries = append(entries, srcEntry{"On-chain Wallet", "onchain", true, breakdown.OnChainSats})
-		}
-		if breakdown.ChannelSats != 0 {
-			entries = append(entries, srcEntry{"Lightning Channels", "channels", false, breakdown.ChannelSats})
-		}
-		if breakdown.ColdStorageSats != 0 {
-			entries = append(entries, srcEntry{"Cold Storage", "cold_storage", false, breakdown.ColdStorageSats})
-		}
-		for _, src := range []string{"strike", "river", "coinbase", "swan"} {
-			if sats, ok := breakdown.ExchangeSats[src]; ok && sats != 0 {
-				entries = append(entries, srcEntry{templateSourceLabel(src), src, true, sats})
-			}
-		}
-
-		for _, e := range entries {
-			pct := 0.0
-			if breakdown.TotalSats > 0 {
-				pct = float64(e.sats) / float64(breakdown.TotalSats) * 100
-			}
-			data.Items = append(data.Items, BreakdownItem{
-				Label:     e.label,
-				SourceKey: e.sourceKey,
-				Clickable: e.clickable,
-				Sats:      e.sats,
-				USD:       satsToUSD(e.sats),
-				Pct:       pct,
-			})
-		}
-		for i := range data.Items {
-			data.DonutColors = append(data.DonutColors, donutChartColors[i%len(donutChartColors)])
-		}
-	}
-
-	// Net flows
-	data.NetFlow = netFlow
-
-	// Cost basis from portfolio position
-	if position != nil {
-		data.TotalCostBasisUSD = position.TotalCostBasisUSD
-		if position.PurchasedSats > 0 {
-			data.AvgCostPerBTC = position.TotalCostBasisUSD / (float64(position.PurchasedSats) / 1e8)
-		}
-		if data.BTCPriceUSD > 0 && data.TotalSats > 0 {
-			data.CurrentValueUSD = float64(data.TotalSats) / 1e8 * data.BTCPriceUSD
-			data.UnrealizedGainUSD = data.CurrentValueUSD - position.TotalCostBasisUSD
-		}
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "portfolio_layout", data); err != nil {
-		h.logger.Printf("failed to render portfolio page: %v", err)
-	}
-}
-
-// portfolioSourceToTxSources maps a donut segment source key to transaction view source values.
-var portfolioSourceToTxSources = map[string][]string{
-	"onchain":  {"lnd_onchain"},
-	"strike":   {"strike"},
-	"river":    {"river"},
-	"coinbase": {"coinbase"},
-	"swan":     {"swan"},
-}
-
-// HandlePortfolioSourceFlows serves GET /portfolio/source-flows — per-source net flows partial.
-func (h *Handler) HandlePortfolioSourceFlows(w http.ResponseWriter, r *http.Request) {
-	sourceKey := r.URL.Query().Get("source")
-	txSources, ok := portfolioSourceToTxSources[sourceKey]
-	if !ok {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		return
-	}
-
-	now := time.Now()
-	period := r.URL.Query().Get("period")
-	var since time.Time
-	switch period {
-	case "90d":
-		since = now.AddDate(0, 0, -90)
-	case "ytd":
-		since = time.Date(now.Year(), 1, 1, 0, 0, 0, 0, time.UTC)
-	case "all":
-		since = time.Time{}
-	default:
-		since = now.AddDate(0, 0, -30)
-	}
-
-	nf, err := h.store.NetFlowSummaryBySource(r.Context(), since, txSources, true)
-	if err != nil {
-		h.logger.Printf("source flows: %v", err)
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if err := h.renderer.Render(w, "source_flows", map[string]interface{}{
-		"SourceKey": sourceKey,
-		"Label":     portfolioSourceLabel(sourceKey),
-		"NetFlow":   nf,
-	}); err != nil {
-		h.logger.Printf("render source flows: %v", err)
-	}
-}
-
-// maskToken shows only the last 4 characters of a token.
-func maskToken(token string) string {
-	if len(token) <= 4 {
-		return "****"
-	}
-	return strings.Repeat("*", 8) + token[len(token)-4:]
-}
