+++ internal/web/templates/tax_layout.html
@@ -7,7 +7,8 @@
   <link rel="stylesheet" href="/static/style.css">
   <script src="/static/htmx.min.js"></script>
 </head>
-<body>
+<body hx-indicator="#htmx-progress">
+  <div id="htmx-progress" class="htmx-indicator"></div>
   <header>
     <div class="container header-inner">
       <div class="header-top">
